chore: move dev module to modules/dev folder
