+++ app/modules/sre/sre.py
@@ -6,9 +6,8 @@ This module contains the main command for the SRE bot. It is responsible for han
 import os
 
 from modules.incident import incident_helper
-from modules import google_service
 from modules.sre import geolocate_helper, webhook_helper
-from modules.dev import aws_dev
+from modules.dev import aws_dev, google_service
 from integrations.slack import commands as slack_commands
 
 help_text = """
@@ -62,13 +61,13 @@ def sre_command(ack, command, logger, respond, client, body):
             respond(f"SRE Bot version: {os.environ.get('GIT_SHA', 'unknown')}")
         case "google":
             if PREFIX == "dev-":
-                google_service.google_service_command(client, body, respond)
+                google_service.google_service_command(ack, client, body, respond)
             else:
                 respond("This command is only available in the dev environment.")
             return
         case "aws":
             if PREFIX == "dev-":
-                aws_dev.aws_dev_command(client, body, respond)
+                aws_dev.aws_dev_command(ack, client, body, respond)
             else:
                 respond("This command is only available in the dev environment.")
             return
